+++ applications/variational/Makefile
@@ -1,16 +1,20 @@
-variational: variational.cc rebalance.o
-	g++ -g -O0 variational.cc rebalance.o -o variational -lm -I/home/frankwillmore/VACUUMMS/include /home/frankwillmore/VACUUMMS/cmake/libraries/libftw_general.so -Wl,-rpath,/home/frankwillmore/VACUUMMS/cmake/libraries/
+# This was the original, now refactoring to 2D etc
+#variational: variational.cc rebalance.o
+#	g++ -g -O0 variational.cc rebalance.o -o variational -lm -I/home/frankwillmore/VACUUMMS/include /home/frankwillmore/VACUUMMS/cmake/libraries/libftw_general.so -Wl,-rpath,/home/frankwillmore/VACUUMMS/cmake/libraries/
 
-variational_2D: variational_2D.cc rebalance.o
-	g++ -g -O0 variational_2D.cc rebalance.o -o variational_2D -lm -I/home/frankwillmore/VACUUMMS/include /home/frankwillmore/VACUUMMS/cmake/libraries/libftw_general.so -Wl,-rpath,/home/frankwillmore/VACUUMMS/cmake/libraries/
+variational_2D.o: variational_2D.cc 
+	g++ -g -O0 -c variational_2D.cc -o variational_2D.o -lm -I/home/frankwillmore/VACUUMMS/include /home/frankwillmore/VACUUMMS/cmake/libraries/libftw_general.so -Wl,-rpath,/home/frankwillmore/VACUUMMS/cmake/libraries/
 
 rebalance.o: rebalance.cc
 	g++ -g -O0 -c rebalance.cc -lm -I/home/frankwillmore/VACUUMMS/include 
 
 # Tests
 
-test_variational: test_variational.cc rebalance.o
-	g++ -g -O0 tester.cc rebalance.o -o tester -lm -I/home/frankwillmore/VACUUMMS/include /home/frankwillmore/VACUUMMS/cmake/libraries/libftw_general.so -Wl,-rpath,/home/frankwillmore/VACUUMMS/cmake/libraries/
+test_variational_2D: test_variational_2D.cc variational_2D.o rebalance.o
+	g++ -g -O0 test_variational_2D.cc variational_2D.o rebalance.o -o test_variational_2D -lm -I/home/frankwillmore/VACUUMMS/include /home/frankwillmore/VACUUMMS/cmake/libraries/libftw_general.so -Wl,-rpath,/home/frankwillmore/VACUUMMS/cmake/libraries/
+
+#test_variational_2D: test_variational_2D.cc rebalance.o
+#	g++ -g -O0 tester.cc rebalance.o -o tester -lm -I/home/frankwillmore/VACUUMMS/include /home/frankwillmore/VACUUMMS/cmake/libraries/libftw_general.so -Wl,-rpath,/home/frankwillmore/VACUUMMS/cmake/libraries/
 
 test_rebalance_1D: test_rebalance_1D.cc rebalance.o
 	g++ test_rebalance_1D.cc rebalance.o -g -O0 -o test_rebalance_1D -lm -I/home/frankwillmore/VACUUMMS/include /home/frankwillmore/VACUUMMS/cmake/libraries/libftw_general.so -Wl,-rpath,/home/frankwillmore/VACUUMMS/cmake/libraries/
@@ -40,4 +44,4 @@ test_rebalance: test_rebalance_1D test_rebalance_2D test_rebalance_3D
 	gnuplot -p -e "plot 'x.dat'" 
 
 clean:
-	rm -f test_rebalance_1D test_rebalance_2D test_rebalance_3D rebalance.o variational
+	rm -f test_rebalance_1D test_rebalance_2D test_rebalance_3D variational *.o
